refactor(reposicao): split AdminReposicaoRevisao (615→115) em hook + componentes
Extrai a lógica (query paginada com caso especial aguardando_fornecedor, memos
de seleção/impacto, mutations approve/update, handlers) para o hook
useRevisaoParametros e quebra o JSX em FiltrosCard, RevisaoTable + SkuRow e
AprovarLoteDialog. Tipos/constantes (PAGE_SIZE, CLASSE_OPTIONS, SkuSugeridoView,
BadgeVariant) em types.ts; reusa helpers de @/lib/reposicao/sku-param e o
SkuDetailSheet já extraído. Comportamento 1:1; +14 testes.
